fix(feature/map-marker-popup) conflicts PR
